+++ src/commonlib/bsd/include/commonlib/bsd/elog.h
@@ -318,7 +318,7 @@ enum cb_err elog_verify_header(const struct elog_header *header);
 const struct event_header *elog_get_next_event(const struct event_header *event);
 const void *event_get_data(const struct event_header *event);
 void elog_fill_timestamp(struct event_header *event, uint8_t sec, uint8_t min,
-			 uint8_t hour, uint8_t mday, uint8_t mon, uint8_t year);
+			 uint8_t hour, uint8_t mday, uint8_t mon, uint16_t year);
 /* Update the checksum at the last byte. */
 void elog_update_checksum(struct event_header *event, uint8_t checksum);
 /* Simple byte checksum for events. */
